move memory operations like copy_to_enclave to enclave_mem.c, refine interfaces
